+++ src/main/java/com/blog/server/blog/dto/UserRequestDto.java
@@ -1,4 +1,4 @@
-package com.blog.server.blog.dto.user;
+package com.blog.server.blog.dto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -6,6 +6,6 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 // Token에 맞춰  규격 정하기
-public class UserInfoDto {
+public class UserRequestDto {
     Long user_id;
 }
